+++ tests/testthat/test-store_boot_def.R
@@ -1,8 +1,3 @@
-skip("WIP")
-# skip_on_cran()
-# skip_if(!interactive(),
-#         message = "standardizedSolution_boot_ci not tested if not interactive")
-
 library(testthat)
 library(semhelpinghands)